requirements.txt
flask
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-api-core==2.19.1
google-auth==2.35.0
google-cloud-core==2.4.1
grpcio==1.65.5
protobuf==4.25.3
Pillow
gradio
google-generativeai